* for 64bit.
git-svn-id: http://svn.sourceforge.jp/svnroot/modchxj/mod_chxj/branches/RELEASE_0_11_0@3189 1a406e8e-add9-4483-a2c8-d8cac5b7c224

+++ src/chxj_dbm.c
@@ -40,12 +40,12 @@ chxj_save_cookie_dbm(request_rec *r, mod_chxj_config *m, const char *cookie_id, 
   apr_dbm_t           *f;
   apr_file_t          *file;
 
-  DBG(r, "REQ[%X] start chxj_save_cookie_dbm() cookie_id:[%s] store_string:[%s]", (apr_size_t)r, cookie_id, store_string);
+  DBG(r, "REQ[%X] start chxj_save_cookie_dbm() cookie_id:[%s] store_string:[%s]", (unsigned int)(apr_size_t)r, cookie_id, store_string);
 
   file = chxj_cookie_db_lock(r);
   if (! file) {
     ERR(r, "mod_chxj: Can't lock cookie db");
-    DBG(r, "REQ[%X] end chxj_save_cookie_dbm() cookie_id:[%s]", (apr_size_t)r, cookie_id);
+    DBG(r, "REQ[%X] end chxj_save_cookie_dbm() cookie_id:[%s]", (unsigned int)(apr_size_t)r, cookie_id);
     return CHXJ_FALSE;
   }
 
@@ -60,7 +60,7 @@ chxj_save_cookie_dbm(request_rec *r, mod_chxj_config *m, const char *cookie_id, 
             "default",
             chxj_cookie_db_name_create(r,m->cookie_db_dir));
     chxj_cookie_db_unlock(r, file);
-    DBG(r, "REQ[%X] end chxj_save_cookie_dbm() cookie_id:[%s]", (apr_size_t)r, cookie_id);
+    DBG(r, "REQ[%X] end chxj_save_cookie_dbm() cookie_id:[%s]", (unsigned int)(apr_size_t)r, cookie_id);
     return CHXJ_FALSE;
   }
 
@@ -83,13 +83,13 @@ chxj_save_cookie_dbm(request_rec *r, mod_chxj_config *m, const char *cookie_id, 
             chxj_cookie_db_name_create(r, m->cookie_db_dir));
     apr_dbm_close(f);
     chxj_cookie_db_unlock(r, file);
-    DBG(r, "REQ[%X] end chxj_save_cookie_dbm() cookie_id:[%s]", (apr_size_t)r, cookie_id);
+    DBG(r, "REQ[%X] end chxj_save_cookie_dbm() cookie_id:[%s]", (unsigned int)(apr_size_t)r, cookie_id);
     return CHXJ_FALSE;
   }
 
   apr_dbm_close(f);
   chxj_cookie_db_unlock(r, file);
-  DBG(r, "REQ[%X] end chxj_save_cookie_dbm() cookie_id:[%s]", (apr_size_t)r, cookie_id);
+  DBG(r, "REQ[%X] end chxj_save_cookie_dbm() cookie_id:[%s]", (unsigned int)(apr_size_t)r, cookie_id);
   return CHXJ_TRUE;
 }
 
@@ -355,11 +355,11 @@ chxj_load_cookie_dbm(request_rec *r, mod_chxj_config *m, const char *cookie_id)
   apr_datum_t             dbmval;
   apr_datum_t             dbmkey;
 
-  DBG(r, "REQ[%X] start chxj_load_cookie_dbm() cookie_id:[%s]", (apr_size_t)r, cookie_id);
+  DBG(r, "REQ[%X] start chxj_load_cookie_dbm() cookie_id:[%s]", (unsigned int)(apr_size_t)r, cookie_id);
   file = chxj_cookie_db_lock(r);
   if (! file) {
     ERR(r, "mod_chxj: Can't lock cookie db");
-    DBG(r, "REQ[%X] end   chxj_load_cookie_dbm() cookie_id:[%s]", (apr_size_t)r, cookie_id);
+    DBG(r, "REQ[%X] end   chxj_load_cookie_dbm() cookie_id:[%s]", (unsigned int)(apr_size_t)r, cookie_id);
     return NULL;
   }
 
@@ -377,7 +377,7 @@ chxj_load_cookie_dbm(request_rec *r, mod_chxj_config *m, const char *cookie_id)
          "default",
          chxj_cookie_db_name_create(r, m->cookie_db_dir));
     chxj_cookie_db_unlock(r, file);
-    DBG(r, "REQ[%X] end   chxj_load_cookie_dbm() cookie_id:[%s]", (apr_size_t)r, cookie_id);
+    DBG(r, "REQ[%X] end   chxj_load_cookie_dbm() cookie_id:[%s]", (unsigned int)(apr_size_t)r, cookie_id);
     return NULL;
   }
 
@@ -398,7 +398,7 @@ chxj_load_cookie_dbm(request_rec *r, mod_chxj_config *m, const char *cookie_id)
            chxj_cookie_db_name_create(r, m->cookie_db_dir));
       apr_dbm_close(f);
       chxj_cookie_db_unlock(r, file);
-      DBG(r, "REQ[%X] end   chxj_load_cookie_dbm() cookie_id:[%s]", (apr_size_t)r, cookie_id);
+      DBG(r, "REQ[%X] end   chxj_load_cookie_dbm() cookie_id:[%s]", (unsigned int)(apr_size_t)r, cookie_id);
       return NULL;
     }
     load_string = apr_palloc(r->pool, dbmval.dsize+1);
@@ -408,7 +408,7 @@ chxj_load_cookie_dbm(request_rec *r, mod_chxj_config *m, const char *cookie_id)
   }
   apr_dbm_close(f);
   chxj_cookie_db_unlock(r, file);
-  DBG(r, "REQ[%X] end   chxj_load_cookie_dbm() cookie_id:[%s] load_string:[%s]", (apr_size_t)r, cookie_id, load_string);
+  DBG(r, "REQ[%X] end   chxj_load_cookie_dbm() cookie_id:[%s] load_string:[%s]", (unsigned int)(apr_size_t)r, cookie_id, load_string);
   return load_string;
 }
